+++ data/overlayCoords.js
@@ -11,5 +11,6 @@ export const overlayCoordsData = new PogObject("FeeshNotifier", {
     "jerryWorkshopTrackerOverlay": { "x": 10, "y": 200, "scale": 1 },
     "wormProfitTrackerOverlay": { "x": 10, "y": 200, "scale": 1 },
     "magmaCoreProfitTrackerOverlay": { "x": 10, "y": 200, "scale": 1 },
+    "abandonedQuarryTrackerOverlay": { "x": 10, "y": 200, "scale": 1 },
     "fishingProfitTrackerOverlay": { "x": 10, "y": 300, "scale": 1 },
 }, 'config/overlayCoords.json');